feat: add MyApplication Link to Java EKS DemoApp (#104)
*Issue #, if available:*  
N/A

*Description of changes:*
This PR creates a myApplication stack, which creates a myApplication
resource and retrieves the `awsApplication` tag value. This tag value is
used to tag SLOs and the EKS cluster, associating all relevant resources
to the myApplication resource.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

*Testing:*
Verified the change by doing the following:
- Deployed successfully. Newest changes from
https://github.com/aws-observability/application-signals-demo/pull/101
fails `cdk synth`