Record invalid date exception explicitly in span (#114)
*Issue #, if available:*

*Description of changes:*

We want to record the invalid date exception explicitly in span.
Otherwise the exception will be captured by the exception handler and
then fail to bubble up in stack to be recorded in span.



By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.